Lots of updates towards eventual 1.0 release
