+++ docs/docs/app/client/index.html
@@ -58,11 +58,7 @@
     <ul class="subnav">
                 <li class="">
                     
-    <a class="" href="../../abstractClient/">AbstractClient</a>
-                </li>
-                <li class="">
-                    
-    <a class="" href="../../httpAbstractClient/">HttpAbstractClient</a>
+    <a class="" href="../../abstractclient/">AbstractClient</a>
                 </li>
                 <li class="">
                     
@@ -92,10 +88,6 @@
                     
     <a class="" href="../../exceptions/missingEncoder/">MissingMessageEncoderException</a>
                 </li>
-                <li class="">
-                    
-    <a class="" href="../../exceptions/api/">APIException</a>
-                </li>
     </ul>
 	    </li>
           
@@ -109,15 +101,27 @@
                 </li>
                 <li class="">
                     
-    <a class="" href="../../device/httpClient/">HttpClient</a>
+    <a class="" href="../../device/parseenvvars/">ParseEnvVar</a>
                 </li>
                 <li class="">
                     
-    <a class="" href="../../device/parseConfigFile/">ParseConfigFile</a>
+    <a class="" href="../../device/parseconfigfile/">ParseConfigFile</a>
                 </li>
                 <li class="">
                     
     <a class="" href="../../device/command/">Command</a>
+                </li>
+                <li class="">
+                    
+    <a class="" href="../../device/managedclient/">ManagedClient</a>
+                </li>
+                <li class="">
+                    
+    <a class="" href="../../device/deviceinfo/">DeviceInfo</a>
+                </li>
+                <li class="">
+                    
+    <a class="" href="../../device/devicefirmware/">DeviceFirmware</a>
                 </li>
     </ul>
 	    </li>
@@ -292,7 +296,7 @@ messaging over MQTT</p>
     <div class="rst-footer-buttons" role="navigation" aria-label="footer navigation">
       
       
-        <a href="../../device/command/" class="btn btn-neutral" title="Command"><span class="icon icon-circle-arrow-left"></span> Previous</a>
+        <a href="../../device/devicefirmware/" class="btn btn-neutral" title="DeviceFirmware"><span class="icon icon-circle-arrow-left"></span> Previous</a>
       
     </div>
   
@@ -318,7 +322,7 @@ messaging over MQTT</p>
     <span class="rst-current-version" data-toggle="rst-current-version">
       
       
-        <span><a href="../../device/command/" style="color: #fcfcfc;">&laquo; Previous</a></span>
+        <span><a href="../../device/devicefirmware/" style="color: #fcfcfc;">&laquo; Previous</a></span>
       
       
     </span>
