+++ docs/docs/device/client/index.html
@@ -58,11 +58,7 @@
     <ul class="subnav">
                 <li class="">
                     
-    <a class="" href="../../abstractClient/">AbstractClient</a>
-                </li>
-                <li class="">
-                    
-    <a class="" href="../../httpAbstractClient/">HttpAbstractClient</a>
+    <a class="" href="../../abstractclient/">AbstractClient</a>
                 </li>
                 <li class="">
                     
@@ -92,10 +88,6 @@
                     
     <a class="" href="../../exceptions/missingEncoder/">MissingMessageEncoderException</a>
                 </li>
-                <li class="">
-                    
-    <a class="" href="../../exceptions/api/">APIException</a>
-                </li>
     </ul>
 	    </li>
           
@@ -112,15 +104,27 @@
                 </li>
                 <li class="">
                     
-    <a class="" href="../httpClient/">HttpClient</a>
+    <a class="" href="../parseenvvars/">ParseEnvVar</a>
                 </li>
                 <li class="">
                     
-    <a class="" href="../parseConfigFile/">ParseConfigFile</a>
+    <a class="" href="../parseconfigfile/">ParseConfigFile</a>
                 </li>
                 <li class="">
                     
     <a class="" href="../command/">Command</a>
+                </li>
+                <li class="">
+                    
+    <a class="" href="../managedclient/">ManagedClient</a>
+                </li>
+                <li class="">
+                    
+    <a class="" href="../deviceinfo/">DeviceInfo</a>
+                </li>
+                <li class="">
+                    
+    <a class="" href="../devicefirmware/">DeviceFirmware</a>
                 </li>
     </ul>
 	    </li>
@@ -174,7 +178,7 @@
               
                 <h1 id="ibmiotf.device.Client">Client</h1>
 
-<pre><code class="python">Client(self, options, logHandlers=None)
+<pre><code class="python">Client(self, config, logHandlers=None)
 </code></pre>
 
 <p>Extends <code>ibmiotf.AbstractClient</code> to implement a device client supporting
@@ -188,12 +192,15 @@ messaging over MQTT</p>
 <p><strong>Configuration Options</strong></p>
 <p>The options parameter expects a Python dictionary containing the following keys:</p>
 <ul>
-<li><code>orgId</code> Your organization ID.</li>
-<li><code>type</code> The type of the device. Think of the device type is analagous to a model number.</li>
-<li><code>id</code> A unique ID to identify a device. Think of the device id as analagous to a serial number.</li>
-<li><code>auth-method</code> The method of authentication. The only method that is currently supported is <code>token</code>.</li>
-<li><code>auth-token</code> An authentication token to securely connect your device to Watson IoT Platform.</li>
-<li><code>clean-session</code> A boolean value indicating whether to use MQTT clean session.</li>
+<li><code>identity.orgId</code> Your organization ID.</li>
+<li><code>identity.typeId</code> The type of the device. Think of the device type is analagous to a model number.</li>
+<li><code>identity.deviceId</code> A unique ID to identify a device. Think of the device id as analagous to a serial number.</li>
+<li><code>auth.token</code> An authentication token to securely connect your device to Watson IoT Platform.</li>
+<li><code>options.domain</code> Optional. A boolean value indicating which Watson IoT Platform domain to connect to (e.g. if you have a dedicated platform instance).</li>
+<li><code>options.mqtt.port</code> Optional. A integer value defining the MQTT port .</li>
+<li><code>options.mqtt.transport</code> Optional. The transport to use for MQTT connectivity - <code>tcp</code> or <code>websockets</code>.</li>
+<li><code>options.mqtt.cleanSession</code> Optional. A boolean value indicating whether to use MQTT clean session.</li>
+<li><code>options.mqtt.caFile</code> Optional. A String value indicating the path to a CA file (in pem format) to use in verifying the server certificate.</li>
 </ul>
 <p><strong>Publishing events</strong></p>
 <p>Events are the mechanism by which devices publish data to the Watson IoT Platform. The device
@@ -270,10 +277,10 @@ on the level of qos used to publish the event:</p>
   
     <div class="rst-footer-buttons" role="navigation" aria-label="footer navigation">
       
-        <a href="../httpClient/" class="btn btn-neutral float-right" title="HttpClient">Next <span class="icon icon-circle-arrow-right"></span></a>
+        <a href="../parseenvvars/" class="btn btn-neutral float-right" title="ParseEnvVar">Next <span class="icon icon-circle-arrow-right"></span></a>
       
       
-        <a href="../../exceptions/api/" class="btn btn-neutral" title="APIException"><span class="icon icon-circle-arrow-left"></span> Previous</a>
+        <a href="../../exceptions/missingEncoder/" class="btn btn-neutral" title="MissingMessageEncoderException"><span class="icon icon-circle-arrow-left"></span> Previous</a>
       
     </div>
   
@@ -299,10 +306,10 @@ on the level of qos used to publish the event:</p>
     <span class="rst-current-version" data-toggle="rst-current-version">
       
       
-        <span><a href="../../exceptions/api/" style="color: #fcfcfc;">&laquo; Previous</a></span>
+        <span><a href="../../exceptions/missingEncoder/" style="color: #fcfcfc;">&laquo; Previous</a></span>
       
       
-        <span style="margin-left: 15px"><a href="../httpClient/" style="color: #fcfcfc">Next &raquo;</a></span>
+        <span style="margin-left: 15px"><a href="../parseenvvars/" style="color: #fcfcfc">Next &raquo;</a></span>
       
     </span>
 </div>
