+++ docs/docs/device/command/index.html
@@ -58,11 +58,7 @@
     <ul class="subnav">
                 <li class="">
                     
-    <a class="" href="../../abstractClient/">AbstractClient</a>
-                </li>
-                <li class="">
-                    
-    <a class="" href="../../httpAbstractClient/">HttpAbstractClient</a>
+    <a class="" href="../../abstractclient/">AbstractClient</a>
                 </li>
                 <li class="">
                     
@@ -92,10 +88,6 @@
                     
     <a class="" href="../../exceptions/missingEncoder/">MissingMessageEncoderException</a>
                 </li>
-                <li class="">
-                    
-    <a class="" href="../../exceptions/api/">APIException</a>
-                </li>
     </ul>
 	    </li>
           
@@ -109,11 +101,11 @@
                 </li>
                 <li class="">
                     
-    <a class="" href="../httpClient/">HttpClient</a>
+    <a class="" href="../parseenvvars/">ParseEnvVar</a>
                 </li>
                 <li class="">
                     
-    <a class="" href="../parseConfigFile/">ParseConfigFile</a>
+    <a class="" href="../parseconfigfile/">ParseConfigFile</a>
                 </li>
                 <li class=" current">
                     
@@ -121,6 +113,18 @@
     <ul class="subnav">
             
     </ul>
+                </li>
+                <li class="">
+                    
+    <a class="" href="../managedclient/">ManagedClient</a>
+                </li>
+                <li class="">
+                    
+    <a class="" href="../deviceinfo/">DeviceInfo</a>
+                </li>
+                <li class="">
+                    
+    <a class="" href="../devicefirmware/">DeviceFirmware</a>
                 </li>
     </ul>
 	    </li>
@@ -203,10 +207,10 @@
   
     <div class="rst-footer-buttons" role="navigation" aria-label="footer navigation">
       
-        <a href="../../app/client/" class="btn btn-neutral float-right" title="Client">Next <span class="icon icon-circle-arrow-right"></span></a>
+        <a href="../managedclient/" class="btn btn-neutral float-right" title="ManagedClient">Next <span class="icon icon-circle-arrow-right"></span></a>
       
       
-        <a href="../parseConfigFile/" class="btn btn-neutral" title="ParseConfigFile"><span class="icon icon-circle-arrow-left"></span> Previous</a>
+        <a href="../parseconfigfile/" class="btn btn-neutral" title="ParseConfigFile"><span class="icon icon-circle-arrow-left"></span> Previous</a>
       
     </div>
   
@@ -232,10 +236,10 @@
     <span class="rst-current-version" data-toggle="rst-current-version">
       
       
-        <span><a href="../parseConfigFile/" style="color: #fcfcfc;">&laquo; Previous</a></span>
+        <span><a href="../parseconfigfile/" style="color: #fcfcfc;">&laquo; Previous</a></span>
       
       
-        <span style="margin-left: 15px"><a href="../../app/client/" style="color: #fcfcfc">Next &raquo;</a></span>
+        <span style="margin-left: 15px"><a href="../managedclient/" style="color: #fcfcfc">Next &raquo;</a></span>
       
     </span>
 </div>
