+++ docs/docs/device/devicefirmware/index.html
@@ -8,7 +8,7 @@
   
   
   <link rel="shortcut icon" href="../../../img/favicon.ico">
-  <title>APIException - IBM Watson IoT Platform Python Library</title>
+  <title>DeviceFirmware - IBM Watson IoT Platform Python Library</title>
   <link href='https://fonts.googleapis.com/css?family=Lato:400,700|Roboto+Slab:400,700|Inconsolata:400,700' rel='stylesheet' type='text/css'>
 
   <link rel="stylesheet" href="../../../css/theme.css" type="text/css" />
@@ -17,9 +17,9 @@
   
   <script>
     // Current page data
-    var mkdocs_page_name = "APIException";
-    var mkdocs_page_input_path = "docs\\exceptions\\api.md";
-    var mkdocs_page_url = "/docs/exceptions/api/";
+    var mkdocs_page_name = "DeviceFirmware";
+    var mkdocs_page_input_path = "docs\\device\\devicefirmware.md";
+    var mkdocs_page_url = "/docs/device/devicefirmware/";
   </script>
   
   <script src="../../../js/jquery-2.1.1.min.js"></script>
@@ -58,11 +58,7 @@
     <ul class="subnav">
                 <li class="">
                     
-    <a class="" href="../../abstractClient/">AbstractClient</a>
-                </li>
-                <li class="">
-                    
-    <a class="" href="../../httpAbstractClient/">HttpAbstractClient</a>
+    <a class="" href="../../abstractclient/">AbstractClient</a>
                 </li>
                 <li class="">
                     
@@ -70,34 +66,27 @@
                 </li>
                 <li class="">
                     
-    <a class="" href="../connect/">ConnectionException</a>
+    <a class="" href="../../exceptions/connect/">ConnectionException</a>
                 </li>
                 <li class="">
                     
-    <a class="" href="../config/">ConfigurationException</a>
+    <a class="" href="../../exceptions/config/">ConfigurationException</a>
                 </li>
                 <li class="">
                     
-    <a class="" href="../auth/">UnsupportedAuthenticationMethod</a>
+    <a class="" href="../../exceptions/auth/">UnsupportedAuthenticationMethod</a>
                 </li>
                 <li class="">
                     
-    <a class="" href="../invalidEvent/">InvalidEventException</a>
+    <a class="" href="../../exceptions/invalidEvent/">InvalidEventException</a>
                 </li>
                 <li class="">
                     
-    <a class="" href="../missingDecoder/">MissingMessageDecoderException</a>
+    <a class="" href="../../exceptions/missingDecoder/">MissingMessageDecoderException</a>
                 </li>
                 <li class="">
                     
-    <a class="" href="../missingEncoder/">MissingMessageEncoderException</a>
-                </li>
-                <li class=" current">
-                    
-    <a class="current" href="./">APIException</a>
-    <ul class="subnav">
-            
-    </ul>
+    <a class="" href="../../exceptions/missingEncoder/">MissingMessageEncoderException</a>
                 </li>
     </ul>
 	    </li>
@@ -108,19 +97,34 @@
     <ul class="subnav">
                 <li class="">
                     
-    <a class="" href="../../device/client/">Client</a>
+    <a class="" href="../client/">Client</a>
                 </li>
                 <li class="">
                     
-    <a class="" href="../../device/httpClient/">HttpClient</a>
+    <a class="" href="../parseenvvars/">ParseEnvVar</a>
                 </li>
                 <li class="">
                     
-    <a class="" href="../../device/parseConfigFile/">ParseConfigFile</a>
+    <a class="" href="../parseconfigfile/">ParseConfigFile</a>
                 </li>
                 <li class="">
                     
-    <a class="" href="../../device/command/">Command</a>
+    <a class="" href="../command/">Command</a>
+                </li>
+                <li class="">
+                    
+    <a class="" href="../managedclient/">ManagedClient</a>
+                </li>
+                <li class="">
+                    
+    <a class="" href="../deviceinfo/">DeviceInfo</a>
+                </li>
+                <li class=" current">
+                    
+    <a class="current" href="./">DeviceFirmware</a>
+    <ul class="subnav">
+            
+    </ul>
                 </li>
     </ul>
 	    </li>
@@ -158,11 +162,11 @@
     
       
         
-          <li>ibmiotf &raquo;</li>
+          <li>ibmiotf.device &raquo;</li>
         
       
     
-    <li>APIException</li>
+    <li>DeviceFirmware</li>
     <li class="wy-breadcrumbs-aside">
       
     </li>
@@ -172,18 +176,10 @@
           <div role="main">
             <div class="section">
               
-                <h1 id="ibmiotf.APIException">APIException</h1>
+                <h1 id="ibmiotf.device.DeviceFirmware">DeviceFirmware</h1>
 
-<pre><code class="python">APIException(self, httpCode, message, response)
+<pre><code class="python">DeviceFirmware(self, version=None, name=None, url=None, verifier=None, state=None, updateStatus=None, updatedDateTime=None)
 </code></pre>
-
-<p>Exception raised when any API call fails</p>
-<p><strong>Attributes</strong></p>
-<ul>
-<li><code>httpCode (int)</code>: The HTTP status code returned</li>
-<li><code>message (string)</code>: The exception message</li>
-<li><code>response (string)</code>: The reponse body that triggered the exception</li>
-</ul>
               
             </div>
           </div>
@@ -191,10 +187,10 @@
   
     <div class="rst-footer-buttons" role="navigation" aria-label="footer navigation">
       
-        <a href="../../device/client/" class="btn btn-neutral float-right" title="Client">Next <span class="icon icon-circle-arrow-right"></span></a>
+        <a href="../../app/client/" class="btn btn-neutral float-right" title="Client">Next <span class="icon icon-circle-arrow-right"></span></a>
       
       
-        <a href="../missingEncoder/" class="btn btn-neutral" title="MissingMessageEncoderException"><span class="icon icon-circle-arrow-left"></span> Previous</a>
+        <a href="../deviceinfo/" class="btn btn-neutral" title="DeviceInfo"><span class="icon icon-circle-arrow-left"></span> Previous</a>
       
     </div>
   
@@ -220,10 +216,10 @@
     <span class="rst-current-version" data-toggle="rst-current-version">
       
       
-        <span><a href="../missingEncoder/" style="color: #fcfcfc;">&laquo; Previous</a></span>
+        <span><a href="../deviceinfo/" style="color: #fcfcfc;">&laquo; Previous</a></span>
       
       
-        <span style="margin-left: 15px"><a href="../../device/client/" style="color: #fcfcfc">Next &raquo;</a></span>
+        <span style="margin-left: 15px"><a href="../../app/client/" style="color: #fcfcfc">Next &raquo;</a></span>
       
     </span>
 </div>
