+++ docs/docs/device/deviceinfo/index.html
@@ -8,7 +8,7 @@
   
   
   <link rel="shortcut icon" href="../../../img/favicon.ico">
-  <title>HttpClient - IBM Watson IoT Platform Python Library</title>
+  <title>DeviceInfo - IBM Watson IoT Platform Python Library</title>
   <link href='https://fonts.googleapis.com/css?family=Lato:400,700|Roboto+Slab:400,700|Inconsolata:400,700' rel='stylesheet' type='text/css'>
 
   <link rel="stylesheet" href="../../../css/theme.css" type="text/css" />
@@ -17,9 +17,9 @@
   
   <script>
     // Current page data
-    var mkdocs_page_name = "HttpClient";
-    var mkdocs_page_input_path = "docs\\device\\httpClient.md";
-    var mkdocs_page_url = "/docs/device/httpClient/";
+    var mkdocs_page_name = "DeviceInfo";
+    var mkdocs_page_input_path = "docs\\device\\deviceinfo.md";
+    var mkdocs_page_url = "/docs/device/deviceinfo/";
   </script>
   
   <script src="../../../js/jquery-2.1.1.min.js"></script>
@@ -58,11 +58,7 @@
     <ul class="subnav">
                 <li class="">
                     
-    <a class="" href="../../abstractClient/">AbstractClient</a>
-                </li>
-                <li class="">
-                    
-    <a class="" href="../../httpAbstractClient/">HttpAbstractClient</a>
+    <a class="" href="../../abstractclient/">AbstractClient</a>
                 </li>
                 <li class="">
                     
@@ -92,10 +88,6 @@
                     
     <a class="" href="../../exceptions/missingEncoder/">MissingMessageEncoderException</a>
                 </li>
-                <li class="">
-                    
-    <a class="" href="../../exceptions/api/">APIException</a>
-                </li>
     </ul>
 	    </li>
           
@@ -107,20 +99,32 @@
                     
     <a class="" href="../client/">Client</a>
                 </li>
+                <li class="">
+                    
+    <a class="" href="../parseenvvars/">ParseEnvVar</a>
+                </li>
+                <li class="">
+                    
+    <a class="" href="../parseconfigfile/">ParseConfigFile</a>
+                </li>
+                <li class="">
+                    
+    <a class="" href="../command/">Command</a>
+                </li>
+                <li class="">
+                    
+    <a class="" href="../managedclient/">ManagedClient</a>
+                </li>
                 <li class=" current">
                     
-    <a class="current" href="./">HttpClient</a>
+    <a class="current" href="./">DeviceInfo</a>
     <ul class="subnav">
             
     </ul>
                 </li>
                 <li class="">
                     
-    <a class="" href="../parseConfigFile/">ParseConfigFile</a>
-                </li>
-                <li class="">
-                    
-    <a class="" href="../command/">Command</a>
+    <a class="" href="../devicefirmware/">DeviceFirmware</a>
                 </li>
     </ul>
 	    </li>
@@ -162,7 +166,7 @@
         
       
     
-    <li>HttpClient</li>
+    <li>DeviceInfo</li>
     <li class="wy-breadcrumbs-aside">
       
     </li>
@@ -172,48 +176,10 @@
           <div role="main">
             <div class="section">
               
-                <h1 id="ibmiotf.device.HttpClient">HttpClient</h1>
+                <h1 id="ibmiotf.device.DeviceInfo">DeviceInfo</h1>
 
-<pre><code class="python">HttpClient(self, options, logHandlers=None)
+<pre><code class="python">DeviceInfo(self)
 </code></pre>
-
-<p>A basic device client with limited capabilies that forgoes
-an active MQTT connection to the service.  Extends <code>ibmiotf.HttpAbstractClient</code>.</p>
-<p><strong>Parameters</strong></p>
-<ul>
-<li><strong>options (dict)</strong>: Configuration options for the client</li>
-<li><strong>logHandlers (list<logging.Handler>)</strong>: Log handlers to configure.  Defaults to <code>None</code>,
-    which will result in a default log handler being created.</li>
-</ul>
-<p><strong>Configuration Options</strong></p>
-<p>The options parameter expects a Python dictionary containing the following keys:</p>
-<ul>
-<li><code>orgId</code> Your organization ID.</li>
-<li><code>type</code> The type of the device. Think of the device type is analagous to a model number.</li>
-<li><code>id</code> A unique ID to identify a device. Think of the device id as analagous to a serial number.</li>
-<li><code>auth-method</code> The method of authentication. The only method that is currently supported is <code>token</code>.</li>
-<li><code>auth-token</code> An authentication token to securely connect your device to Watson IoT Platform.</li>
-</ul>
-<p>The HTTP client supports four content-types for posted events:</p>
-<ul>
-<li><code>application/xml</code>: for events/commands using message format <code>xml</code></li>
-<li><code>text/plain; charset=utf-8</code>: for events/commands using message format <code>plain</code></li>
-<li><code>application/octet-stream</code>: for events/commands using message format <code>bin</code></li>
-<li><code>application/json</code>: the default for all other message formats.</li>
-</ul>
-<h2 id="ibmiotf.device.HttpClient.publishEvent">publishEvent</h2>
-
-<pre><code class="python">HttpClient.publishEvent(self, event, msgFormat, data)
-</code></pre>
-
-<p>Publish an event over HTTP(s) as given supported format</p>
-<p><strong>Raises</strong></p>
-<ul>
-<li><code>MissingMessageEncoderException</code>: If there is no registered encoder for <code>msgFormat</code></li>
-<li><code>Exception</code>: If something went wrong</li>
-</ul>
-<p><strong>Returns</strong></p>
-<p><code>int</code>: The HTTP status code for the publish</p>
               
             </div>
           </div>
@@ -221,10 +187,10 @@ an active MQTT connection to the service.  Extends <code>ibmiotf.HttpAbstractCli
   
     <div class="rst-footer-buttons" role="navigation" aria-label="footer navigation">
       
-        <a href="../parseConfigFile/" class="btn btn-neutral float-right" title="ParseConfigFile">Next <span class="icon icon-circle-arrow-right"></span></a>
+        <a href="../devicefirmware/" class="btn btn-neutral float-right" title="DeviceFirmware">Next <span class="icon icon-circle-arrow-right"></span></a>
       
       
-        <a href="../client/" class="btn btn-neutral" title="Client"><span class="icon icon-circle-arrow-left"></span> Previous</a>
+        <a href="../managedclient/" class="btn btn-neutral" title="ManagedClient"><span class="icon icon-circle-arrow-left"></span> Previous</a>
       
     </div>
   
@@ -250,10 +216,10 @@ an active MQTT connection to the service.  Extends <code>ibmiotf.HttpAbstractCli
     <span class="rst-current-version" data-toggle="rst-current-version">
       
       
-        <span><a href="../client/" style="color: #fcfcfc;">&laquo; Previous</a></span>
+        <span><a href="../managedclient/" style="color: #fcfcfc;">&laquo; Previous</a></span>
       
       
-        <span style="margin-left: 15px"><a href="../parseConfigFile/" style="color: #fcfcfc">Next &raquo;</a></span>
+        <span style="margin-left: 15px"><a href="../devicefirmware/" style="color: #fcfcfc">Next &raquo;</a></span>
       
     </span>
 </div>
