+++ docs/docs/exceptions/invalidEvent/index.html
@@ -58,11 +58,7 @@
     <ul class="subnav">
                 <li class="">
                     
-    <a class="" href="../../abstractClient/">AbstractClient</a>
-                </li>
-                <li class="">
-                    
-    <a class="" href="../../httpAbstractClient/">HttpAbstractClient</a>
+    <a class="" href="../../abstractclient/">AbstractClient</a>
                 </li>
                 <li class="">
                     
@@ -95,10 +91,6 @@
                     
     <a class="" href="../missingEncoder/">MissingMessageEncoderException</a>
                 </li>
-                <li class="">
-                    
-    <a class="" href="../api/">APIException</a>
-                </li>
     </ul>
 	    </li>
           
@@ -112,15 +104,27 @@
                 </li>
                 <li class="">
                     
-    <a class="" href="../../device/httpClient/">HttpClient</a>
+    <a class="" href="../../device/parseenvvars/">ParseEnvVar</a>
                 </li>
                 <li class="">
                     
-    <a class="" href="../../device/parseConfigFile/">ParseConfigFile</a>
+    <a class="" href="../../device/parseconfigfile/">ParseConfigFile</a>
                 </li>
                 <li class="">
                     
     <a class="" href="../../device/command/">Command</a>
+                </li>
+                <li class="">
+                    
+    <a class="" href="../../device/managedclient/">ManagedClient</a>
+                </li>
+                <li class="">
+                    
+    <a class="" href="../../device/deviceinfo/">DeviceInfo</a>
+                </li>
+                <li class="">
+                    
+    <a class="" href="../../device/devicefirmware/">DeviceFirmware</a>
                 </li>
     </ul>
 	    </li>
