+++ docs/docs/exceptions/missingEncoder/index.html
@@ -58,11 +58,7 @@
     <ul class="subnav">
                 <li class="">
                     
-    <a class="" href="../../abstractClient/">AbstractClient</a>
-                </li>
-                <li class="">
-                    
-    <a class="" href="../../httpAbstractClient/">HttpAbstractClient</a>
+    <a class="" href="../../abstractclient/">AbstractClient</a>
                 </li>
                 <li class="">
                     
@@ -95,10 +91,6 @@
             
     </ul>
                 </li>
-                <li class="">
-                    
-    <a class="" href="../api/">APIException</a>
-                </li>
     </ul>
 	    </li>
           
@@ -112,15 +104,27 @@
                 </li>
                 <li class="">
                     
-    <a class="" href="../../device/httpClient/">HttpClient</a>
+    <a class="" href="../../device/parseenvvars/">ParseEnvVar</a>
                 </li>
                 <li class="">
                     
-    <a class="" href="../../device/parseConfigFile/">ParseConfigFile</a>
+    <a class="" href="../../device/parseconfigfile/">ParseConfigFile</a>
                 </li>
                 <li class="">
                     
     <a class="" href="../../device/command/">Command</a>
+                </li>
+                <li class="">
+                    
+    <a class="" href="../../device/managedclient/">ManagedClient</a>
+                </li>
+                <li class="">
+                    
+    <a class="" href="../../device/deviceinfo/">DeviceInfo</a>
+                </li>
+                <li class="">
+                    
+    <a class="" href="../../device/devicefirmware/">DeviceFirmware</a>
                 </li>
     </ul>
 	    </li>
@@ -189,7 +193,7 @@
   
     <div class="rst-footer-buttons" role="navigation" aria-label="footer navigation">
       
-        <a href="../api/" class="btn btn-neutral float-right" title="APIException">Next <span class="icon icon-circle-arrow-right"></span></a>
+        <a href="../../device/client/" class="btn btn-neutral float-right" title="Client">Next <span class="icon icon-circle-arrow-right"></span></a>
       
       
         <a href="../missingDecoder/" class="btn btn-neutral" title="MissingMessageDecoderException"><span class="icon icon-circle-arrow-left"></span> Previous</a>
@@ -221,7 +225,7 @@
         <span><a href="../missingDecoder/" style="color: #fcfcfc;">&laquo; Previous</a></span>
       
       
-        <span style="margin-left: 15px"><a href="../api/" style="color: #fcfcfc">Next &raquo;</a></span>
+        <span style="margin-left: 15px"><a href="../../device/client/" style="color: #fcfcfc">Next &raquo;</a></span>
       
     </span>
 </div>
