+++ docs/index.html
@@ -80,11 +80,7 @@
     <ul class="subnav">
                 <li class="">
                     
-    <a class="" href="docs/abstractClient/">AbstractClient</a>
-                </li>
-                <li class="">
-                    
-    <a class="" href="docs/httpAbstractClient/">HttpAbstractClient</a>
+    <a class="" href="docs/abstractclient/">AbstractClient</a>
                 </li>
                 <li class="">
                     
@@ -114,10 +110,6 @@
                     
     <a class="" href="docs/exceptions/missingEncoder/">MissingMessageEncoderException</a>
                 </li>
-                <li class="">
-                    
-    <a class="" href="docs/exceptions/api/">APIException</a>
-                </li>
     </ul>
 	    </li>
           
@@ -131,15 +123,27 @@
                 </li>
                 <li class="">
                     
-    <a class="" href="docs/device/httpClient/">HttpClient</a>
+    <a class="" href="docs/device/parseenvvars/">ParseEnvVar</a>
                 </li>
                 <li class="">
                     
-    <a class="" href="docs/device/parseConfigFile/">ParseConfigFile</a>
+    <a class="" href="docs/device/parseconfigfile/">ParseConfigFile</a>
                 </li>
                 <li class="">
                     
     <a class="" href="docs/device/command/">Command</a>
+                </li>
+                <li class="">
+                    
+    <a class="" href="docs/device/managedclient/">ManagedClient</a>
+                </li>
+                <li class="">
+                    
+    <a class="" href="docs/device/deviceinfo/">DeviceInfo</a>
+                </li>
+                <li class="">
+                    
+    <a class="" href="docs/device/devicefirmware/">DeviceFirmware</a>
                 </li>
     </ul>
 	    </li>
@@ -188,10 +192,14 @@
             <div class="section">
               
                 <h1 id="python-for-ibm-watson-iot-platform">Python for IBM Watson IoT Platform</h1>
+<p><a href="https://travis-ci.org/ibm-watson-iot/iot-python"><img alt="Build Status" src="https://travis-ci.org/ibm-watson-iot/iot-python.svg?branch=master" /></a>
+<a href="https://github.com/ibm-watson-iot/iot-python/issues"><img alt="GitHub issues" src="https://img.shields.io/github/issues/ibm-watson-iot/iot-python.svg" /></a>
+<a href="https://github.com/ibm-watson-iot/iot-python/blob/master/LICENSE"><img alt="GitHub" src="https://img.shields.io/github/license/ibm-watson-iot/iot-python.svg" /></a>
+<a href="https://pypi.org/project/ibmiotf/"><img alt="PyPI" src="https://img.shields.io/pypi/v/ibmiotf.svg" /></a></p>
 <p>Python module for interacting with the <a href="https://internetofthings.ibmcloud.com">IBM Watson IoT Platform</a>.</p>
 <ul>
-<li><a href="https://www.python.org/downloads/release/python-360/">Python 3.6</a></li>
-<li><a href="https://www.python.org/downloads/release/python-2713/">Python 2.7</a></li>
+<li><a href="https://www.python.org/downloads/release/python-370/">Python 3.7</a></li>
+<li><a href="https://www.python.org/downloads/release/python-2715/">Python 2.7</a></li>
 </ul>
 <p>Note: Support for MQTT with TLS requires at least Python v2.7.9 or v3.4, and openssl v1.0.1</p>
 <h2 id="dependencies">Dependencies</h2>
@@ -201,8 +209,6 @@
 <li><a href="https://pypi.python.org/pypi/pytz">pytz</a></li>
 <li><a href="https://pypi.python.org/pypi/requests">requests</a></li>
 <li><a href="https://pypi.python.org/pypi/requests_toolbelt">requests_toolbelt</a></li>
-<li><a href="https://pypi.python.org/pypi/dicttoxml">dicttoxml</a></li>
-<li><a href="https://pypi.python.org/pypi/xmltodict">xmltodict</a></li>
 </ul>
 <h2 id="installation">Installation</h2>
 <p>Install the latest version of the library with pip</p>
@@ -235,11 +241,11 @@
 <li><strong>Auto Reconnect</strong>: All clients support automatic reconnect to the Platform in the event of a network interruption.</li>
 <li><strong>Event/Command publish</strong>: Offers a pure HTTP client supporting messaging over HTTP in addition to the full features client that utilizes HTTP and MQTT technologies as appropriate </li>
 <li><strong>Data Format Support</strong>: JSON, XML, &amp; user-defined.</li>
+<li><strong>Websockets</strong>: Support device/gateway/application connectivity to Watson IoT Platform using WebSocket</li>
 </ul>
 <h2 id="unsupported-features">Unsupported Features</h2>
 <ul>
 <li><strong>Client side Certificate based authentication</strong>: <a href="https://console.ng.bluemix.net/docs/services/IoT/reference/security/RM_security.html">Client side Certificate based authentication</a>n</li>
-<li><strong>Websockets</strong>: Support device/gateway/application connectivity to Watson IoT Platform using WebSocket</li>
 </ul>
               
             </div>
@@ -248,7 +254,7 @@
   
     <div class="rst-footer-buttons" role="navigation" aria-label="footer navigation">
       
-        <a href="docs/abstractClient/" class="btn btn-neutral float-right" title="AbstractClient">Next <span class="icon icon-circle-arrow-right"></span></a>
+        <a href="docs/abstractclient/" class="btn btn-neutral float-right" title="AbstractClient">Next <span class="icon icon-circle-arrow-right"></span></a>
       
       
     </div>
@@ -276,7 +282,7 @@
       
       
       
-        <span style="margin-left: 15px"><a href="docs/abstractClient/" style="color: #fcfcfc">Next &raquo;</a></span>
+        <span style="margin-left: 15px"><a href="docs/abstractclient/" style="color: #fcfcfc">Next &raquo;</a></span>
       
     </span>
 </div>
@@ -290,5 +296,5 @@
 
 <!--
 MkDocs version : 0.17.3
-Build Date UTC : 2018-07-30 15:27:06
+Build Date UTC : 2018-12-05 02:14:33
 -->
