+++ docs/search.html
@@ -51,11 +51,7 @@
     <ul class="subnav">
                 <li class="">
                     
-    <a class="" href="docs/abstractClient/">AbstractClient</a>
-                </li>
-                <li class="">
-                    
-    <a class="" href="docs/httpAbstractClient/">HttpAbstractClient</a>
+    <a class="" href="docs/abstractclient/">AbstractClient</a>
                 </li>
                 <li class="">
                     
@@ -85,10 +81,6 @@
                     
     <a class="" href="docs/exceptions/missingEncoder/">MissingMessageEncoderException</a>
                 </li>
-                <li class="">
-                    
-    <a class="" href="docs/exceptions/api/">APIException</a>
-                </li>
     </ul>
 	    </li>
           
@@ -102,15 +94,27 @@
                 </li>
                 <li class="">
                     
-    <a class="" href="docs/device/httpClient/">HttpClient</a>
+    <a class="" href="docs/device/parseenvvars/">ParseEnvVar</a>
                 </li>
                 <li class="">
                     
-    <a class="" href="docs/device/parseConfigFile/">ParseConfigFile</a>
+    <a class="" href="docs/device/parseconfigfile/">ParseConfigFile</a>
                 </li>
                 <li class="">
                     
     <a class="" href="docs/device/command/">Command</a>
+                </li>
+                <li class="">
+                    
+    <a class="" href="docs/device/managedclient/">ManagedClient</a>
+                </li>
+                <li class="">
+                    
+    <a class="" href="docs/device/deviceinfo/">DeviceInfo</a>
+                </li>
+                <li class="">
+                    
+    <a class="" href="docs/device/devicefirmware/">DeviceFirmware</a>
                 </li>
     </ul>
 	    </li>
