+++ docs/sitemap.xml
@@ -4,7 +4,7 @@
     
     <url>
      <loc>/</loc>
-     <lastmod>2018-07-30</lastmod>
+     <lastmod>2018-12-05</lastmod>
      <changefreq>daily</changefreq>
     </url>
     
@@ -12,62 +12,50 @@
     
         
     <url>
-     <loc>/docs/abstractClient/</loc>
-     <lastmod>2018-07-30</lastmod>
-     <changefreq>daily</changefreq>
-    </url>
-        
-    <url>
-     <loc>/docs/httpAbstractClient/</loc>
-     <lastmod>2018-07-30</lastmod>
+     <loc>/docs/abstractclient/</loc>
+     <lastmod>2018-12-05</lastmod>
      <changefreq>daily</changefreq>
     </url>
         
     <url>
      <loc>/docs/message/</loc>
-     <lastmod>2018-07-30</lastmod>
+     <lastmod>2018-12-05</lastmod>
      <changefreq>daily</changefreq>
     </url>
         
     <url>
      <loc>/docs/exceptions/connect/</loc>
-     <lastmod>2018-07-30</lastmod>
+     <lastmod>2018-12-05</lastmod>
      <changefreq>daily</changefreq>
     </url>
         
     <url>
      <loc>/docs/exceptions/config/</loc>
-     <lastmod>2018-07-30</lastmod>
+     <lastmod>2018-12-05</lastmod>
      <changefreq>daily</changefreq>
     </url>
         
     <url>
      <loc>/docs/exceptions/auth/</loc>
-     <lastmod>2018-07-30</lastmod>
+     <lastmod>2018-12-05</lastmod>
      <changefreq>daily</changefreq>
     </url>
         
     <url>
      <loc>/docs/exceptions/invalidEvent/</loc>
-     <lastmod>2018-07-30</lastmod>
+     <lastmod>2018-12-05</lastmod>
      <changefreq>daily</changefreq>
     </url>
         
     <url>
      <loc>/docs/exceptions/missingDecoder/</loc>
-     <lastmod>2018-07-30</lastmod>
+     <lastmod>2018-12-05</lastmod>
      <changefreq>daily</changefreq>
     </url>
         
     <url>
      <loc>/docs/exceptions/missingEncoder/</loc>
-     <lastmod>2018-07-30</lastmod>
-     <changefreq>daily</changefreq>
-    </url>
-        
-    <url>
-     <loc>/docs/exceptions/api/</loc>
-     <lastmod>2018-07-30</lastmod>
+     <lastmod>2018-12-05</lastmod>
      <changefreq>daily</changefreq>
     </url>
         
@@ -77,25 +65,43 @@
         
     <url>
      <loc>/docs/device/client/</loc>
-     <lastmod>2018-07-30</lastmod>
+     <lastmod>2018-12-05</lastmod>
      <changefreq>daily</changefreq>
     </url>
         
     <url>
-     <loc>/docs/device/httpClient/</loc>
-     <lastmod>2018-07-30</lastmod>
+     <loc>/docs/device/parseenvvars/</loc>
+     <lastmod>2018-12-05</lastmod>
      <changefreq>daily</changefreq>
     </url>
         
     <url>
-     <loc>/docs/device/parseConfigFile/</loc>
-     <lastmod>2018-07-30</lastmod>
+     <loc>/docs/device/parseconfigfile/</loc>
+     <lastmod>2018-12-05</lastmod>
      <changefreq>daily</changefreq>
     </url>
         
     <url>
      <loc>/docs/device/command/</loc>
-     <lastmod>2018-07-30</lastmod>
+     <lastmod>2018-12-05</lastmod>
+     <changefreq>daily</changefreq>
+    </url>
+        
+    <url>
+     <loc>/docs/device/managedclient/</loc>
+     <lastmod>2018-12-05</lastmod>
+     <changefreq>daily</changefreq>
+    </url>
+        
+    <url>
+     <loc>/docs/device/deviceinfo/</loc>
+     <lastmod>2018-12-05</lastmod>
+     <changefreq>daily</changefreq>
+    </url>
+        
+    <url>
+     <loc>/docs/device/devicefirmware/</loc>
+     <lastmod>2018-12-05</lastmod>
      <changefreq>daily</changefreq>
     </url>
         
@@ -105,7 +111,7 @@
         
     <url>
      <loc>/docs/app/client/</loc>
-     <lastmod>2018-07-30</lastmod>
+     <lastmod>2018-12-05</lastmod>
      <changefreq>daily</changefreq>
     </url>
