+++ src/ibmiotf/__init__.py
@@ -412,137 +412,3 @@ class MissingMessageEncoderException(Exception):
 
     def __str__(self):
         return "No message encoder defined for message format: %s" % self.format
-
-
-class APIException(Exception):
-    """
-    Exception raised when any API call fails
-    
-    # Attributes
-    httpCode (int): The HTTP status code returned
-    message (string): The exception message
-    response (string): The reponse body that triggered the exception
-    """
-    def __init__(self, httpCode, message, response):
-        self.httpCode = httpCode
-        self.message = message
-        self.response = response
-
-    def __str__(self):
-        if self.response:
-            return "[%s] %s . Response: %s" % (self.httpCode, self.message, self.response)
-        return "[%s] %s" % (self.httpCode, self.message)
-
-
-class HttpAbstractClient(object):
-    """
-    The underlying client object utilised for Platform connectivity 
-    over HTPP in devices, gateways, and applications.
-    
-    Restricted to HTTP only.  Unless for some technical reason
-    you are unable to use the full MQTT-enable client there really 
-    is no need to use this alternative feature-limited client as 
-    installing this library means you already have access to the 
-    rich MQTT/HTTP client implementation.
-    
-    The HTTP client supports four content-types for posted events:
-    
-    - `application/xml`: for events/commands using message format `xml`
-    - `text/plain; charset=utf-8`: for events/commands using message format `plain`
-    - `application/octet-stream`: for events/commands using message format `bin`
-    - `application/json`: the default for all other message formats.
-    """
-    def __init__(self, clientId, logHandlers=None):
-        # Configure logging
-        self.logger = logging.getLogger(self.__module__+"."+self.__class__.__name__)
-        self.logger.setLevel(logging.INFO)
-
-        # Remove any existing log handlers we may have picked up from getLogger()
-        self.logger.handlers = []
-
-        if logHandlers:
-            if isinstance(logHandlers, list):
-                # Add all supplied log handlers
-                for handler in logHandlers:
-                    self.logger.addHandler(handler)
-            else:
-                # Add the supplied log handler
-                self.logger.addHandler(logHandlers)
-        else:
-            # Generate a default rotating file log handler and stream handler
-            logFileName = '%s.log' % (clientId.replace(":", "_"))
-            fhFormatter = logging.Formatter('%(asctime)-25s %(name)-25s ' + ' %(levelname)-7s %(message)s')
-            rfh = RotatingFileHandler(logFileName, mode='a', maxBytes=1024000 , backupCount=0, encoding=None, delay=True)
-            rfh.setFormatter(fhFormatter)
-
-            ch = logging.StreamHandler()
-            ch.setFormatter(fhFormatter)
-            ch.setLevel(logging.DEBUG)
-
-            self.logger.addHandler(rfh)
-            self.logger.addHandler(ch)
-
-        # Initialize default message encoders and decoders.
-        self._messageEncoderModules = {}
-
-    def connect(self):
-        """
-        Connect is a no-op with HTTP-only client, but the presence of this method makes it easy 
-        to switch between using HTTP & MQTT client implementation
-        """
-        pass
-
-    def disconnect(self):
-        """
-        Disconnect is a no-op with HTTP-only client, but the presence of this method makes it easy 
-        to switch between using HTTP & MQTT client implementation
-        """
-        pass
-
-    def getMessageEncoderModule(self, messageFormat):
-        """
-        Get the Python module that is currently defined as the encoder/decoder for a specified message format.
-        
-        # Arguments
-        messageFormat (string): The message format to retrieve the encoder for
-        
-        # Returns
-        Boolean: The python module, or `None` if there is no codec defined for the `messageFormat`
-        """
-        return self._messageEncoderModules[messageFormat]
-
-    def setMessageEncoderModule(self, messageFormat, module):
-        """
-        Set a Python module as the encoder/decoder for a specified message format.
-        
-        # Arguments
-        messageFormat (string): The message format to retreive the encoder for
-        module (module): The Python module to set as the encoder/decoder for `messageFormat`
-        """
-        self._messageEncoderModules[messageFormat] = module
-
-    def _logAndRaiseException(self, e):
-        """
-        Logs an exception at log level `critical` before raising it.
-        
-        # Arguments
-        e (Exception): The exception to log/raise
-        """
-        self.logger.critical(str(e))
-        raise e
-
-    def _getContentType(self, dataFormat):
-        """
-        Determines the content type for the HTTP message
-        """
-        # Default content type is json
-        contentType = "application/json"
-        if dataFormat == "text":
-            contentType = "text/plain; charset=utf-8"
-        elif dataFormat == "xml":
-            contentType = "application/xml"
-        elif dataFormat == "bin":
-            contentType = "application/octet-stream"
-        
-        # Return derived content type
-        return contentType
