+++ src/ibmiotf/device/__init__.py
@@ -1,0 +1,17 @@
+# *****************************************************************************
+# Copyright (c) 2014, 2018 IBM Corporation and other Contributors.
+#
+# All rights reserved. This program and the accompanying materials
+# are made available under the terms of the Eclipse Public License v1.0
+# which accompanies this distribution, and is available at
+# http://www.eclipse.org/legal/epl-v10.html
+# *****************************************************************************
+
+
+# Expose public API for this package
+from ibmiotf.device.client import Client
+from ibmiotf.device.command import Command
+from ibmiotf.device.config import ParseConfigFile, ParseEnvVars
+from ibmiotf.device.deviceFirmware import DeviceFirmware
+from ibmiotf.device.deviceInfo import DeviceInfo
+from ibmiotf.device.managedClient import ManagedClient
