+++ test/test_device_cfg.py
@@ -1,0 +1,69 @@
+# *****************************************************************************
+# Copyright (c) 2016 IBM Corporation and other Contributors.
+#
+# All rights reserved. This program and the accompanying materials
+# are made available under the terms of the Eclipse Public License v1.0
+# which accompanies this distribution, and is available at
+# http://www.eclipse.org/legal/epl-v10.html
+#
+# *****************************************************************************
+
+from nose.tools import *
+import ibmiotf.device
+from ibmiotf import ConfigurationException
+import testUtils
+
+class TestDeviceCfg(testUtils.AbstractTest):
+
+    def testMissingOptions(self):
+        with assert_raises(ConfigurationException) as e:
+            ibmiotf.device.Client({})
+        assert_equal(e.exception.reason, 'Missing identity from configuration')
+
+    def testMissingOrg(self):
+        with assert_raises(ConfigurationException) as e:
+            ibmiotf.device.Client({
+                "identity": {
+                    "orgId": None, "typeId": "myType", "deviceId": "myId"
+                },
+                "auth": { "token" : "myToken" }
+            })
+        assert_equal(e.exception.reason, 'Missing identity.orgId from configuration')
+
+    def testMissingType(self):
+        with assert_raises(ConfigurationException) as e:
+            ibmiotf.device.Client({
+                "identity": {
+                    "orgId": "myOrg", "typeId": None, "deviceId": "myId"
+                },
+                "auth": { "token" : "myToken" }
+            })
+        assert_equal(e.exception.reason, 'Missing identity.typeId from configuration')
+
+    def testMissingId(self):
+        with assert_raises(ConfigurationException) as e:
+            ibmiotf.device.Client({
+                "identity": {
+                    "orgId": "myOrg", "typeId": "myType", "deviceId": None
+                },
+                "auth": { "token" : "myToken" }
+            })
+        assert_equal(e.exception.reason, 'Missing identity.deviceId from configuration')
+
+    def testMissingAuthToken(self):
+        with assert_raises(ConfigurationException) as e:
+            ibmiotf.device.Client({
+                "identity": {
+                    "orgId": "myOrg", "typeId": "myType", "deviceId": "myId"
+                },
+                "auth": { "token" : None }
+            })
+            ibmiotf.device.Client({"org": self.ORG_ID, "type": self.registeredDevice.typeId, "id": self.registeredDevice.deviceId,
+                                   "auth-method": None, "auth-token": self.registeredDevice.authToken})
+        assert_equal(e.exception.reason, 'Missing auth.token from configuration')
+
+    def testMissingConfigFile(self):
+        deviceFile="InvalidFile.out"
+        with assert_raises(ConfigurationException) as e:
+            ibmiotf.device.ParseConfigFile(deviceFile)
+        assert_equal(e.exception.reason, "Error reading device configuration file 'InvalidFile.out' (No such file or directory)")
